third_party/WebKit/LayoutTests/fast/events/message-event-max-ports.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Excessively long port sequences should gracefully fail.');

let ports = [];
ports.length = 1<<29;
shouldThrow('new MessageEvent("message", {ports:ports})');
</script>
</body>
</html>
